.github/requirements-docs.txt
mkdocs>=1.6,<2
mkdocs-material>=9.5
mkdocs-glightbox>=0.4
Jinja2>=3.1
PyYAML>=6.0
